pyproject.toml
[project]
name = "notes-app"
version = "0.1.0"
description = ""
authors = [
    {name = "je_biere",email = "lenkis.fedor@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.115.13,<0.116.0",
  "uvicorn[standard]>=0.34.3,<0.35.0",
  "sqlalchemy[asyncio] (>=2.0.41,<3.0.0)",
  "alembic (>=1.16.2,<2.0.0)",
  "pydantic[email] (>=2.11.7,<3.0.0)",
  "asyncpg (>=0.30.0,<0.31.0)",
  "passlib[bcrypt] (>=1.7.4,<2.0.0)",
  "pyjwt[cryptography] (>=2.10.1,<3.0.0)",
  "pydantic-settings (>=2.9.1,<3.0.0)",
  "cryptography (>=45.0.4,<46.0.0)"
]

[tool.poetry]
packages = [
  { include = "src", from = "." }
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"


[tool.poetry.group.migration.dependencies]
psycopg2-binary = "^2.9.10"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
